requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Vector Database
chromadb==0.4.18

# Cache
redis==5.0.1
hiredis==2.2.3

# HTTP Client
httpx==0.25.2

# LLM Providers
google-generativeai==0.3.2
openai==1.3.7  # For DeepSeek (OpenAI-compatible)

# Sentiment Analysis
transformers==4.35.2
torch==2.1.1
sentencepiece==0.1.99

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Monitoring & Logging
prometheus-client==0.19.0
structlog==23.2.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Code Quality
black==23.11.0
flake8==6.1.0
mypy==1.7.1



